fix speed control

+++ firmware/motor.h
@@ -10,7 +10,7 @@ extern "C" {
 
 void motor_config();
 void motor_enable(int enable);
-void motor_set_speed(int8_t speedLeft, int8_t speedRight);
+void motor_set_speed(int16_t speedLeft, int16_t speedRight);
 void servo_tilt_set(uint8_t val);
 void servo_tilt_stop();
